Move scroll-to-hash method into module
